Hud & radar corrections
Now the hud and radar works fine. The hud do not interrogate the property tree anymore and is faster.
The radar can work without writing a tree, is improved and cleaner. The target object is in a seprate file.
The hud loop is done on the main loop, which has been cleaned.
There is still lot of work to do.