Revised aero - wing aerofoils and CG
I came to the concluse that the wing aerofoils that I was using must be wrong.

As the only reference I can find is on "The Incomplete Guide to Airfoil Usage" (below) I've amended to use that and shifted the CG forwards to 7.5 which gives a better looking pitch moment (CMM1).

Dassault Mirage III                                    ? 3.5%                        ? 3.5%
Dassault Mirage IV                                     ? 3.8%                        ? 3.2%
Dassault Mirage V                                      ? 3.5%                        ? 3.5%

So now we are using NACA-0004 (3.5%) and NACA-0003 (3.4%);

CG was tested by moving forwards and backwards until a sensible looking CMM1 plot was produced. VSP thinks the CG should be at 7.898; 25% MAC is at 8.1; currently using 7.5

This is now flyable - although still twitchy and unstable.