Full detail run that has
1. reasonable pitch moment
2. controllable up to 30 degrees alpha by having sufficient authority when both elevons are fully deflected
3. reasonable basic yaw moment.

However this still isn't right; but it is in the right area. More work to do on the CoG; as it must be in front of the main gear by a decent margin. I've measured this as being forwards of 7.96m based on a 10degree angle of rotation on takeoff.