Cleaning the HUD
Cleaning the code and cleaning useless information when we have a selected target. The Alt and kts are now displayed on canvas. Need to add mach, and H (agl), then put a box around the acceleration figures that are present on take off. the HUD will become more and more accurate