Aero changes
Revised geometry based on wingspan and wing area described here https://faq-fra.aviatechno.net/guide/m2000.php

Changed wing incidence to -4 degrees on the basis that the pitch moment curve is had 0 at -ve alpha which probably is wrong. Previously the incidence was 1.4 degrees. This value could be wrong - as yet I've not found a definitive source and measuring the airframe is somewhat inconclusive.